exam-scheduler/requirements.txt
ortools==9.7.2996
openai==0.27.8
pandas==2.2.3
python-dotenv==1.0.1
pymongo==4.10.1
Flask==3.1.0
Flask-PyMongo==2.3.0
Flask-Cors==5.0.0